# Conflicts:
#	ProSuite/src/ProSuite.DomainModel.Core/QA/SupportedInstanceDescriptors.cs

+++ src/ProSuite.AGP.WorkList/Datasource/WorkItemTable.cs
@@ -131,7 +131,7 @@ namespace ProSuite.AGP.WorkList.Datasource
 			return false;
 		}
 
-		public override int GetNativeRowCount()
+		public override long GetNativeRowCount()
 		{
 			throw new NotSupportedException();
 		}
